pyproject.toml
[project]
name = "order-management-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Jakub Wilczek",email = "jacob_0909@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "order_management_api", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
